requirements.txt
ccxt==4.1.77
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
asyncio-mqtt==0.16.1
MetaTrader5==5.0.45
schedule==1.2.0
